replace depracated init method of GCKDeviceScanner

+++ Pod/Classes/AUMediaLibrary.m
@@ -49,7 +49,7 @@
         if ([downloadTask.response isKindOfClass:[NSHTTPURLResponse class]]) {
             NSUInteger statusCode = [(NSHTTPURLResponse *)downloadTask.response statusCode];
             if (!(statusCode >= 200 && statusCode < 300)) {
-                NSLog(@"%@ failed (statusCode = %ld)", [downloadTask.originalRequest.URL lastPathComponent], statusCode);
+                NSLog(@"%@ failed (statusCode = %ld)", [downloadTask.originalRequest.URL lastPathComponent], (long)statusCode);
                 return nil;
             }
         }
